ctf.core: Support field class alias in CTF2 Metadata
Allow field-class-alias fragments in root metadata node.

Add new class FieldClass to be used instead of JsonObject. Create and
register a FieldClassDeserializer that will accept a JSON Object as a
field class or a JSON String as a field class alias and lookup a
previously occurring field class alias with that name.

Support a string as a field class alias in TypeAliasParser and lookup a
previously occurring field class alias with that name in the Metadata
root.

[Added] Support field class alias in CTF2 Metadata

Change-Id: I7a7cf52a1ce36ba726ba2ab568ed6e9b9cfc4803
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/204823
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>