releng: Fix update site for simrel
Add org.apache.commons.lang3 to update site, as this plugin is not found
in default platform or dependent features simrel update sites.

Revert SLF4J bundle names to those of current platform.

Revert default releng files to be for the current platform, and use
staging folder for 4.30 platform version of releng files.

Change-Id: I86dc1527992807ef25873b95032389c7a7717adb
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/205227
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>