java-doc: Exclude o.e.tc.tmf.remote from java-doc due to build errors
After upgrading tycho to 4.0.x and maven 3.9.x building of java-doc
fails with o.e.remote not found for the o.e.tc.tmf.remote plug-ins.

With this the java-doc for all Trace Compass API but TMF remote can be
built and deployed again.

Upgrade maven java doc plugin to 3.6.2.

Change-Id: If1a02bed8da8b1b12f5397fac28836277faf929c
Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/205616
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>