tmf: Add data type to other event aspects
Based on the getDataType method in the ITmfEventAspect interfact, the
aspects that are implementing this interface now have their own data
type. The data type is STRING by default, hence, the aspects that their
data type is not considered as STRING are now equipped with the correct
data type.

[Added] Proper corresponding data type to the aspects that their default
data type is not considered as STRING

Signed-off-by: Kaveh Shahedi <kaveh.shahedi@ericsson.com>