Add tree structures to statistics to add more hierarchies
The SegmentStoreStatisticsDataProvider previously supported
only two layers of hierarchy in the statistics. This commit
introduces a tree structure extending the IStatistics interface,
allowing for the addition of more hierarchical layers to the
entries. Additionally, the SegmentStoreStatisticsDataProvider
has been refactored into an abstract class to share common code
between it and the new SegmentStoreTreeStatisticsDataProvider.

[Added] tree structures to statistics to more hierarchies.

Signed-off-by: Siwei Zhang <siwei.zhang@ericsson.com>