Add default values to SegmentStoreStatisticsAspects
The statistics previously used default values like
Double.MAX_VALUE and Double.MIN_VALUE for fields such
as min and max, which were displayed when no elements
were present in the statistics. This commit introduces
a default value of null for situations where there are
no elements in the statistics.

[Added] Default value for SegmentStoreStatisticsAspects

Signed-off-by: Siwei Zhang <siwei.zhang@ericsson.com>