tmf: Add API to add and remove analysis modules to/from ITmfTrace
Many analysis utilities require analysis modules to be present in the
ITmfTrace.getAnalysisModules() collection. This one is filled using the
TMF analysis framework (see IAnalysisModuleSource implementation).
The analysis framework will populate available IAnalysisModuesHelpers at
startup (plug-in defined) or when a refresh is requested (XML analysis).
For that a central storage for the modules configuration needs to be
available to be parsed which are available for all ITmfTrace extensions.
If a trace is opened a ITmfTrace instance is created and the applicable
analysis module is instantiated and stored inside the ITmfTrace object.

Right now it's not possible to update available analysis module helpers
in the TMF analysis framework based on a configuration specific for
a trace instance.

This patch allows to add analysis modules to ITmfTrace and by-pass the
TMF analysis framework and IAnalysisModuleSource extension point to
provide custom analysis.

Note that those analysis modules won't show up in the Project Explorer
and hence it's meant for the trace server use case.

[Added] API to add and remove analysis modules to/from ITmfTrace

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>