tmf: Introduce a JSON configuration class for de-/serialization
It writes the JSON parameter string as JSON object to file and
deserializes it back to a JSON string.

It provide conversion methods to/from ITmfConfiguration. It only
uses the JSON parameter String and not parameter Map<String, Object>.

For parameter in format Map<String, Object> the user has to provide
a custom serialization method instead.

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>