tmf: Add configuration setter/getter in IAnalysisModule
This is meant to be used when creating and managing analysis modules
programmatically and not through the TMF Analysis Framework (e.g.
defining the analysis in the plug-in.xml). The configuration has to be
set before the analysis is scheduled, which is not supported by the TMF
Analysis Framework.

This is useful in the trace server context for configurable analysis.

[Added] configuration setter/getter in IAnalysisModule

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>