tmf: add capabilities data structure to IDataProviderDescriptor
With this a data provider factory instance can advertise special
capabilities that can be applied for the given data provider.

This commit adds code for that and implements the "canCreate" and
"canDelete" capability. "canCreate" indicates that this instance
can create a derived data provider. Use the IDataProviderFactory method
getAdapter(ITmfDataProviderConfigurator.class) to obtains an instance of
ITmfDataProviderConfigurator. This interface has methods to get
List<ITmfConfigurationSourceType> describing the parameters needed to
create a derived data provider using method
ITmfDataProviderConfigurator#createDataProvider.

"canDelete" indicates a the data provider can be deleted using
method ITmfDataProviderConfigurator#deleteDataProviderDescriptor.

[Added] capabilities data structure to IDataProviderDescriptor

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>