tmf: Add auto-expand level in TmfTreeModel
The auto-expand level is supposed to be used for the input of the tree.
The value 0 means that there is no auto-expand; 1 means that top-level
elements are expanded, but not their children; 2 means that top-level
elements are expanded, and their children, but not grand-children; and
so on.

The value -1 means that all subtrees should be expanded (default).

[Added] Add auto-expand level in TmfTreeModel

Signed-off-by: Hriday Panchasara hriday.panchasara@ericsson.com
Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>