linux: set auto-expand-level in ResourcesStatusDataProvider
Setting the auto-expand-level to 1 of the returned TmfTreeModel when
calling fetchTree() will indicate to the clients that only the top-level
should be expanded and their children supposed to be collapsed.

With this trace server clients can apply thhis and it fixes FE issue
https://github.com/eclipse-cdt-cloud/theia-trace-extension/issues/801

Note: The classic Eclipse UI has been doing this in the UI code without
that setting (see ResourcesView).

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>