linux.core: : Add IRQ handlers instead of generic IRQs
Currently, the Resources view only shows IRQs as a single entity.
This makes it harder to distinguish which handler is actually running,
especially in cases where multiple IRQ handlers are mapped to the same
lane.

This change updates the view to display IRQ handlers instead of just
the IRQ, providing finer granularity and improving readability when a
single lane contains multiple handlers.

Benefits:
* Clearer visualization of interrupt activity.
* Easier to identify which handler is active at a given time.
* Better distinction when multiple handlers share one IRQ lane.

[Added] Add handlers to IRQs in resources views

Change-Id: I131790ec6dad3b28328abea0c5d84b280a7b196e
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>