ctf: introduce Dynamic Length String datatype
It's a mix of a sequence and a string.

A dynamic-length string field class is an abstract string field class which describes dynamic-length string fields.

A static-length string field is a sequence of zero or more contiguous encoded Unicode codepoints. All the encoded codepoints of a dynamic-length string field before the first “NULL” (U+0000) codepoint, if any, form the resulting string value. The first U+0000 codepoint, if any, and all the following bytes are considered padding (garbage data).

Change-Id: I19bbef5eb2913d8252e469e477363da7eecacdcb
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>